Contribute Gradle plugin
Contributing the [Gradle plugin](https://github.com/ben-manes/gradle-jooq-plugin) required,
 - Updating to the package names
 - Adding to the Maven build. The Gradle plugin dependencies piggyback
   on Flyway's repository, as Gradle does not officially publish the
   dependencies for building plugins in another build tool.
 - Assigning copyright

This contribution is not tested due to a broader build failure in
`FlashbackTable`.

Note that this plugin took a simple and fast approach by leveraging
Groovy's XML markeup syntax. This provided a concise extension, at
the cost of flexibility to customize the plugin in multiple build
scripts (e.g. subproject adding data type converters). A richer
plugin should be created, but exceeded the scope of what was needed
at the time of development.