[#4554] Ensure batches are sent
Due to overlap in the method prototypes of `DSLContext#batch(Query)` and
`DSLContext#batch(Query...)` calls intended to invoke the later with a
single argument actually invoke the former. To solve this we check if
`BatchBindStep#bind` was subsequently called and, if not, proceed as if
the var args version was called.