Support Postgres Plus Advanced Server (PPAS)
Add pg_catalog routines and tables in order to support codegen of PPAS
packages.
* edb_package - packages metadata
* pg_namespace - additional fields to allow for joins to pg_proc
* pg_proc - additional fields for joining to pg_namespace
* row_number - used to count the number of overloads

Add PPAS Database, Package and Routine definitions.

PostgresPlusDatabaseDefinition - extends the PostgresDatabaseDefinition
so that the code generation methods from the superclass can be used
where they conform to PPAS. Includes further predicate to prevent the
creation of triggers as routines (triggers should rarely, if ever, be
used as stand-alone functions).

PostgresPlusPackageDefinition - extends the AbstractPackageDefinition.
There is no PostgresPackageDefinition to extend.

PostgresPlusRoutineDefinition - extends the AbstractRoutineDefinition.
This is required because the PostgresRoutineDefinition hides the
AbstractRoutineDefinition constructor, making it impossible to pass the
package definition, which is necessary to hook the new class up to the
plumbing of the code generator for emitting packages.