[jOOQ/jOOQ#8736] Add Tools#tryParseInt() and Tools#tryParseLong()
For the benefit of performance Tools#renderAndBind() now no longer uses
Integer#parseInt() when deciding whether a placeholder is a keyword or
an index of a bind value.

Also AbstractQuery#bind(String, Object) and DefaultBinding#parse(Class,
String) no longer rely on Integer#parseInt() or Long#parseLong().

The new methods in Tools are based on the Guava method
Longs#tryParse(String, int).