[jOOQ/jOOQ#8806] Correctly parse identifiers that look like keywords
Parser doesn't correctly parse identifiers that can conflict with keywords. For example, when parsing TRIM(TRAILING.C) where TRAILING is a table reference, we parse it as a keyword and don't check if it is really just a qualified identifier. Whenever there's a . following a keyword, we'll treat the token as a non-keyword.