[jOOQ/jOOQ#8986] New DSLContext#meta(Source...) API
Adds a new internal `DSLContext#meta(Source...)` API which can be used to construct a `Meta` object for a list of DDL scripts (similarly to the existing `DDLDatabase`), but for other use cases such as the `Parser`.

This new method uses the new internal `Source` abstraction to represent a DDL script, which can be constructed using any of the existing static `of()` factory methods or optionally by a custom implementation class.