[jOOQ/jOOQ#9157] [jOOQ/jOOR#57] Try again
When running the Java 8 distribution's code of jOOR on Java 9, CACHED_LOOKUP_CONSTRUCTOR is null because we're not allowed to access that from JDK 9+. This means we can no longer use it in reflection logic, e.g. when calling default methods. We have to use the Java 9+ logic via meta-reflection.