[jOOQ/jOOQ#9764] Rework formatting of SQL
One big formatting problem is in QueryPartList. 1) it adds unnecessary
whitespace at the end of lines, because it hardcodes a ", " separator.
That is easy to fix. Also, it needs to decide whether at the beginning
of a list, an extra separator is required. That's something that a
container QueryPart needs to decide for the QueryPartList, whereas the
QueryPartList gets to decide if that separator is a newline or
whitespace.

A few other formatting issues are fixed with this one, including:

- [jOOQ/jOOQ#1962] Formatting of SQL rendered by window functions
- [jOOQ/jOOQ#3479] Awkward formatting of "short" functions
- [jOOQ/jOOQ#5488] Excess newline in formatted SELECT w/out FROM clause
- [jOOQ/jOOQ#9764] This issue