[jOOQ/jOOQ#11700] Add DSL.using(io.r2dbc.spi.Connection)
It should be possible to wrap a io.r2dbc.spi.Connection in a single-connection R2DBC ConnectionFactory, which produces Connection proxies that never close the underlying Connection. This works the same way as using DSL.using(Connection) which is wrapped in a JDBC ConnectionProvider that never closes the underlying Connection.