[jOOQ/jOOQ#11700] Add resourceful, blocking DSL.using() support.
The resourceful DSL.using(String) et al. methods are useful for quick-and-dirty testing things. They offer a blocking way to connect to a driver, returning a CloseableDSLContext, whose close() method is also blocking. While that isn't 100% in the true spirit of R2DBC, we can still offer the functionality as users might use this for quick tests.