[jOOQ/jOOQ#14356] Fixed regression
The goal of this algorithm is to iteratively traverse the tree breadth
first, keeping a heap based stack to emulate depth first traversal. The
previous algorithm mixed up the order on that stack, such that some
trees were traversed in the wrong order, producing wrong results, e.g.

  (1 + 2) + (3 + 4)

Was rendered as

  (3 + 1 + 2 + 4)

# Conflicts:
#jOOQ/src/main/java/org/jooq/impl/Expression.java