[jOOQ/jOOQ#14513] Support column COMMENT clause in ALTER TABLE .. ADD
This implements the feature for dialects with native support.

Before implementing it for other dialects using an anonymous block, let's refactor the internals, first. There are 2 modes of adding columns to a table, both work differently, even if they're pretty much the same in terms of the QOM, even if the syntax may differ.