[jOOQ/jOOQ#14694] Re-wrap actual Val, don't create a ConvertedVal
The initial fix attempted to create a ConvertedVal from the original Val, but that didn't always produce the required bind value casts, e.g. in Firebird. When jOOQ's internals create a Val (as opposed to the user), then the interesting value is really only the value, not the wrapper, so we can discard the temporary wrapper as soon as we have explicit type information.

As an aside, it is particularly important not to apply this logic e.g. when parsing / translating SQL, in case of which we currently still rely on Val identity for caching (!)