[jOOQ/jOOQ#2682] [jOOQ/jOOQ#15632] Support set operations (UNION etc.)
This includes a refactoring where ScopeStackElement is subclassed for specialisation. We need one key per type of ScopeStackElement, such that values can be null for certain keys. That way, the new ScopeDefinerScopeStack element can be absent even if the DataKeyScopeStackElement is present on all scopes. Future refactors might further specialise the DefaultScopeStackElement, which contains the query part specific contents (e.g. join tree)