[jOOQ/jOOQ#2682] [jOOQ/jOOQ#15632] Handle UPDATE .. SET
UPDATE .. SET must not produce rows that will no longer be accessible under any policy, so for UPDATE, a policy might generate 2 predicates. A first one for the WHERE part to allow only updates on rows that already are valid according the policy, and a second one to check if the SET clause will produce rows that will still be valid.