[jOOQ/jOOQ#15755] Make RenderImplicitJoinType.THROW default for to-many
To-many implicit joins are a controversial feature for the reasons
discussed in various issues. THROW is now the default for them, to make
sure no RDBMS with weird implementation for the SCALAR_SUBQUERY
implementation confuses users.

This change also correctly generates implicit joins from paths declared
in FROM, in case sub-paths are skipped.