[jOOQ/jOOQ#9506] Fix mixing SCRIPT with SCHEMA
A SCRIPT is allowed to modify a schema that is otherwise managed by SCHEMA files. There must be a SCHEMA file documenting the outcome. Other than for that purpose, that SCHEMA file must be ignored, and no diff must be applied from the previously known SCHEMA file.

This currently only works if the SCHEMA and SCRIPT files are contained in the same Commit. It's currently not possible to fix a mistake, where the SCHEMA file has been forgotten, other than to edit the SCRIPT file.