[jOOQ/jOOQ#9506] Improve Migration.baseline() behaviour
It should be possible to specify a specific Commit as the baseline. e.g. with v2, so we want to be able to assume v2 was correctly installed in previous migrations / with other products, etc. By default, the baseline is set to the current commit (which is often not what we want, but a reasonable default).