[jOOQ/jOOQ#18682] [jOOQ/jOOQ#18686] Work around API generation limitation
We currently cannot have pro-only clauses of a OSS available statement, as the QOM API will still expose the pro-only properties for arguments. This is mostly a cosmetic limitation as the DSL API is not directly reachable, and the QOM API simply has no effect because the implementation is missing.